requirements.txt
Django==2.2
jinja2
django-picklefield
django-import-export
psycopg2-binary==2.7.4

celery==4.2.1
redis==2.10.6

numpy>=1.9.0
pandas>=0.23.0
networkx>=2.1
scipy
openpyxl
xlrd
matplotlib==2.2.3
matplotlib-venn
bioservices
pathos
appdirs
plotly==2.7
seaborn
wordcloud
sortedcontainers
statsmodels

#py2cytoscape
pydot
pydotplus
IPython>=5.3.0
jupyter==1.0.0
magine==0.0.16
